+++ src/ZoomHub/Config.hs
@@ -4,41 +4,22 @@
 
 module ZoomHub.Config
   ( Config (..),
-    RackspaceConfig,
     defaultPort,
-    raxApiKey,
-    raxContainer,
-    raxContainerPath,
-    raxUsername,
-
-    -- * Process content status
-    ProcessContent (..),
-    parseProcessContent,
   )
 where
 
-import Data.Aeson ((.=), ToJSON, Value (String), object, toJSON)
+import Data.Aeson ((.=), ToJSON, object, toJSON)
 import qualified Data.ByteString.Lazy as BL
-import qualified Data.Text as T
 import Data.Time.Units (Second)
+import Data.Time.Units.Instances ()
 import qualified Database.PostgreSQL.Simple as PGS
 import Database.PostgreSQL.Simple.Instances ()
-import GHC.Generics (Generic)
-import Network.URI (URI, parseRelativeReference)
 import Network.URI.Instances ()
 import Network.Wai (Middleware)
 import Squeal.PostgreSQL.Pool (Pool)
-import System.Envy
-  ( DefConfig,
-    FromEnv,
-    Option (..),
-    customPrefix,
-    defConfig,
-    dropPrefixCount,
-    fromEnv,
-    gFromEnvCustom,
-  )
-import ZoomHub.Rackspace.CloudFiles (Container, parseContainer)
+import qualified ZoomHub.Config.AWS as AWS
+import ZoomHub.Config.ProcessContent (ProcessContent (..))
+import ZoomHub.Config.Uploads (Uploads (..))
 import ZoomHub.Storage.PostgreSQL (Connection)
 import ZoomHub.Types.BaseURI (BaseURI)
 import ZoomHub.Types.ContentBaseURI (ContentBaseURI)
@@ -50,7 +31,8 @@ defaultPort = 8000
 
 data Config
   = Config
-      { baseURI :: BaseURI,
+      { aws :: AWS.Config,
+        baseURI :: BaseURI,
         contentBaseURI :: ContentBaseURI,
         dbConnInfo :: PGS.ConnectInfo,
         dbConnPool :: Pool Connection,
@@ -62,8 +44,8 @@ data Config
         openSeadragonScript :: String,
         port :: Integer,
         processContent :: ProcessContent,
+        uploads :: Uploads,
         publicPath :: FilePath,
-        rackspace :: RackspaceConfig,
         staticBaseURI :: StaticBaseURI,
         tempPath :: TempPath,
         version :: String
@@ -78,55 +60,11 @@ instance ToJSON Config where
         "dbConnPoolIdleTime" .= dbConnPoolIdleTime,
         "dbConnPoolMaxResourcesPerStripe" .= dbConnPoolMaxResourcesPerStripe,
         "dbConnPoolNumStripes" .= dbConnPoolNumStripes,
-        "processContent" .= processContent,
         "port" .= port,
+        "processContent" .= processContent,
         "publicPath" .= publicPath,
         "staticBaseURI" .= staticBaseURI,
         "tempPath" .= tempPath,
+        "uploads" .= uploads,
         "version" .= version
       ]
-
--- Rackspace
-data RackspaceConfig
-  = RackspaceConfig
-      { raxUsername :: String, -- RACKSPACE_USERNAME
-        raxApiKey :: String, -- RACKSPACE_API_KEY
-        raxContainer :: Container, -- RACKSPACE_CONTAINER
-        raxContainerPath :: URI -- RACKSPACE_CONTAINER_PATH
-      }
-  deriving (Generic, Show)
-
--- Default configuration will be used for fields that could not be
--- retrieved from the environment:
-instance DefConfig RackspaceConfig where
-  defConfig = RackspaceConfig
-    { raxUsername = "zoomingservice",
-      raxApiKey = error "Missing `raxApiKey`",
-      raxContainer = case parseContainer "cache-development" of
-        Just container -> container
-        _ -> error "Failed to parse `raxContainer`.",
-      raxContainerPath = case parseRelativeReference "content" of
-        Just containerPath -> containerPath
-        _ -> error "Failed to parse `raxContainerPath`."
-    }
-
-instance FromEnv RackspaceConfig where
-  fromEnv = gFromEnvCustom Option
-    { dropPrefixCount = 3,
-      customPrefix = "RACKSPACE"
-    }
-
--- ProcessContent
-data ProcessContent
-  = ProcessNoContent
-  | ProcessExistingContent
-  | ProcessExistingAndNewContent
-  deriving (Eq, Show)
-
-parseProcessContent :: String -> ProcessContent
-parseProcessContent "ProcessExistingContent" = ProcessExistingContent
-parseProcessContent "ProcessExistingAndNewContent" = ProcessExistingAndNewContent
-parseProcessContent _ = ProcessNoContent
-
-instance ToJSON ProcessContent where
-  toJSON = String . T.pack . show